Add basic deformed shape viewer.
The API for this is kinda wonky and probably needs another pass. I'd
like to add support for (1) changing the scale without regenerating the
model and (2) scrubbing through deformation time history. (1) seems
easier to implement, but looks to be troublesome. (2) requires further
changes to the data structure, which will take some thought.